fix null array, add rowindex

+++ src/Constants.cs
@@ -2,7 +2,7 @@
 
 public static class Constants
 {
-    public static readonly long Null = BitConverter.ToInt32([0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE]);
+    public static readonly long Null = BitConverter.ToInt64([0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE, 0xFE]);
     public static readonly byte[] StringNullTerminator = [0, 0, 0, 0];
 }
